[BC-Breaking] Kill experimental descriptor APIs
This completely removes all outdated API surfaces for TMA:
- `tl._experimental_descriptor_load`
- `tl._experimental_descriptor_store`
- `tl._experimental_reinterpret_tensor_descriptor`
- `triton.tools.experimental_descriptor.create_{1,2}d_tma_descriptor`
- `triton.tools.experimental_descriptor.TmaDescKernelParam`
- `tl.extra.cuda.experiemental_device_tensormap_create1d`
- `tl.extra.cuda.experiemental_device_tensormap_create2d`
- `tl.extra.cuda.experiemental_tensormap_fenceproxy_acquire`

I also rename `triton.tools.experimental_descriptor` ->
`triton.tools.tensor_descriptor` to reflect that it's
no longer experimental.
