+++ python/triton/language/__init__.py
@@ -27,12 +27,9 @@ from .standard import (
 from .core import (
     PropagateNan,
     TRITON_MAX_TENSOR_NUMEL,
-    _experimental_descriptor_load,
-    _experimental_descriptor_store,
     load_tensor_descriptor,
     store_tensor_descriptor,
     make_tensor_descriptor,
-    _experimental_reinterpret_tensor_descriptor,
     tensor_descriptor,
     tensor_descriptor_type,
     add,
@@ -134,12 +131,9 @@ from .random import (
 __all__ = [
     "PropagateNan",
     "TRITON_MAX_TENSOR_NUMEL",
-    "_experimental_descriptor_load",
-    "_experimental_descriptor_store",
     "load_tensor_descriptor",
     "store_tensor_descriptor",
     "make_tensor_descriptor",
-    "_experimental_reinterpret_tensor_descriptor",
     "tensor_descriptor",
     "abs",
     "add",
